// tmp/test/models/PubRolePageDateRes.ts
/**
 * CNet API
 * CNet基础框架API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { PubRole } from '../models/PubRole';
import { ResCode } from '../models/ResCode';
import { HttpFile } from '../http/http';

export class PubRolePageDateRes {
    'code'?: ResCode;
    'msg'?: string | null;
    'count'?: number;
    'totalPage'?: number;
    'data'?: Array<PubRole> | null;
    'pageNum'?: number;
    'pageSize'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "code",
            "baseName": "code",
            "type": "ResCode",
            "format": ""
        },
        {
            "name": "msg",
            "baseName": "msg",
            "type": "string",
            "format": ""
        },
        {
            "name": "count",
            "baseName": "count",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "totalPage",
            "baseName": "totalPage",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<PubRole>",
            "format": ""
        },
        {
            "name": "pageNum",
            "baseName": "pageNum",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "pageSize",
            "baseName": "pageSize",
            "type": "number",
            "format": "int32"
        }    ];

    static getAttributeTypeMap() {
        return PubRolePageDateRes.attributeTypeMap;
    }

    public constructor() {
    }
}